Move TimeComparator to own file
