+++ src/main/kotlin/org/amshove/kluent/DateTime.kt
@@ -45,89 +45,33 @@ infix fun LocalDate.shouldBeIn(theMonth: Month) = this `should be in` theMonth
 infix fun LocalDate.`should be in year`(theYear: Int) = assertTrue("Expected $this to be in $theYear, but was ${this.year}", this.year == theYear)
 infix fun LocalDate.shouldBeInYear(theYear: Int) = this `should be in year` theYear
 
-fun Int.hours() = TimeComparator(addedHours = this)
-fun Int.minutes() = TimeComparator(addedMinutes = this)
-fun Int.seconds() = TimeComparator(addedSeconds = this)
+fun Int.hours() = LocalTimeComparator(addedHours = this)
+fun Int.minutes() = LocalTimeComparator(addedMinutes = this)
+fun Int.seconds() = LocalTimeComparator(addedSeconds = this)
 
-infix fun LocalTime.`should be`(timeComparator: TimeComparator): TimeComparator {
-    timeComparator.startTime = this
-    return timeComparator
+infix fun LocalTime.`should be`(localTimeComparator: LocalTimeComparator): LocalTimeComparator {
+    localTimeComparator.startTime = this
+    return localTimeComparator
 }
-infix fun LocalTime.shouldBe(timeComparator: TimeComparator) = this `should be` timeComparator
+infix fun LocalTime.shouldBe(localTimeComparator: LocalTimeComparator) = this `should be` localTimeComparator
 
-infix fun LocalTime.`should be at least`(timeComparator: TimeComparator): TimeComparator {
-    timeComparator.startTime = this
-    timeComparator.timeComparatorType = TimeComparatorType.AtLeast
-    return timeComparator
+infix fun LocalTime.`should be at least`(localTimeComparator: LocalTimeComparator): LocalTimeComparator {
+    localTimeComparator.startTime = this
+    localTimeComparator.timeComparatorType = TimeComparatorType.AtLeast
+    return localTimeComparator
 }
-infix fun LocalTime.shouldBeAtLeast(timeComparator: TimeComparator) = this `should be at least` timeComparator
+infix fun LocalTime.shouldBeAtLeast(localTimeComparator: LocalTimeComparator) = this `should be at least` localTimeComparator
 
-infix fun LocalTime.`should be at most`(timeComparator: TimeComparator): TimeComparator {
-    timeComparator.startTime = this
-    timeComparator.timeComparatorType = TimeComparatorType.AtMost
-    return timeComparator
+infix fun LocalTime.`should be at most`(localTimeComparator: LocalTimeComparator): LocalTimeComparator {
+    localTimeComparator.startTime = this
+    localTimeComparator.timeComparatorType = TimeComparatorType.AtMost
+    return localTimeComparator
 }
-infix fun LocalTime.shouldBeAtMost(timeComparator: TimeComparator) = this `should be at most` timeComparator
+infix fun LocalTime.shouldBeAtMost(localTimeComparator: LocalTimeComparator) = this `should be at most` localTimeComparator
 
 
-infix fun TimeComparator.after(theOther: LocalTime) = this.assertAfter(theOther)
-infix fun TimeComparator.before(theOther: LocalTime) = this.assertBefore(theOther)
-
-class TimeComparator(internal val addedHours: Int = 0, internal val addedMinutes: Int = 0, internal val addedSeconds: Int = 0) {
-    internal lateinit var startTime: LocalTime
-    internal var timeComparatorType = TimeComparatorType.Exactly
-
-    internal fun assertAfter(theOther: LocalTime) =
-            when (timeComparatorType) {
-                TimeComparatorType.AtLeast -> assertAtLeastAfter(theOther)
-                TimeComparatorType.Exactly -> assertExactlyAfter(theOther)
-                TimeComparatorType.AtMost -> assertAtMostAfter(theOther)
-            }
-
-
-    internal fun assertBefore(theOther: LocalTime) =
-            when (timeComparatorType) {
-                TimeComparatorType.AtLeast -> assertAtLeastBefore(theOther)
-                TimeComparatorType.Exactly -> assertExactlyBefore(theOther)
-                TimeComparatorType.AtMost -> assertAtMostBefore(theOther)
-            }
-
-    private fun assertAtLeastAfter(theOther: LocalTime) {
-        val comparedTime = calculateComparedTime(theOther)
-        assertTrue("Expected $startTime to be at least { $addedHours hours, $addedMinutes minutes, $addedSeconds seconds } after $theOther", startTime >= comparedTime)
-    }
-
-    private fun assertAtMostAfter(theOther: LocalTime) {
-        val comparedTime = calculateComparedTime(theOther)
-        assertTrue("Expected $startTime to be at most { $addedHours hours, $addedMinutes minutes, $addedSeconds seconds } after $theOther", startTime <= comparedTime)
-    }
-
-    private fun assertExactlyAfter(theOther: LocalTime) {
-        val comparedTime = calculateComparedTime(theOther)
-        assertTrue("Expected $startTime to be { $addedHours hours, $addedMinutes minutes, $addedSeconds seconds } after $theOther", startTime == comparedTime)
-    }
-
-    private fun assertExactlyBefore(theOther: LocalTime) {
-        val comparedTime = calculateComparedTime(theOther, -1)
-        assertTrue("Expected $startTime to be { $addedHours hours, $addedMinutes minutes, $addedSeconds seconds } before $theOther", startTime == comparedTime)
-    }
-
-    private fun assertAtLeastBefore(theOther: LocalTime) {
-        val comparedTime = calculateComparedTime(theOther, -1)
-        assertTrue("Expected $startTime to be at least { $addedHours hours, $addedMinutes minutes, $addedSeconds seconds } before $theOther", startTime <= comparedTime)
-    }
-
-    private fun assertAtMostBefore(theOther: LocalTime) {
-        val comparedTime = calculateComparedTime(theOther, -1)
-        assertTrue("Expected $startTime to be at most { $addedHours hours, $addedMinutes minutes, $addedSeconds seconds } before $theOther", startTime >= comparedTime)
-    }
-
-    private fun calculateComparedTime(time: LocalTime, multiplier: Int = 1) =
-            time.plusHours(addedHours.toLong() * multiplier)
-                    .plusMinutes(addedMinutes.toLong() * multiplier)
-                    .plusSeconds(addedSeconds.toLong() * multiplier)
-
-}
+infix fun LocalTimeComparator.after(theOther: LocalTime) = this.assertAfter(theOther)
+infix fun LocalTimeComparator.before(theOther: LocalTime) = this.assertBefore(theOther)
 
 internal enum class TimeComparatorType {
     AtMost,
